+++ src/main/java/roomescape/reservationTime/ReservationTime.java
@@ -7,16 +7,16 @@ public class ReservationTime {
     private Long id;
     private String startAt;
 
+    public ReservationTime(Long id) {
+        this.id = id;
+    }
+
     public Long getId() {
         return id;
     }
 
     public String getStartAt() {
         return startAt;
-    }
-
-    public ReservationTime(Long id) {
-        this.id = id;
     }
 
     public ReservationTime(String startAt) {
