+++ src/main/java/roomescape/reservationTime/ReservationTimeResponseDto.java
@@ -5,13 +5,17 @@ public class ReservationTimeResponseDto {
     private Long id;
     private String startAt;
 
-    public Long getId() { return id; }
-
-    public String getStartAt() { return startAt; }
-
     public ReservationTimeResponseDto(Long id, String startAt) {
         this.id = id;
         this.startAt = startAt;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getStartAt() {
+        return startAt;
     }
 
     @Override
