+++ src/test/java/roomescape/reservation/ReservationRepositoryTest.java
@@ -7,8 +7,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.dao.DataAccessException;
 import org.springframework.jdbc.core.JdbcTemplate;
-import roomescape.time.ReservationTime;
-import roomescape.time.ReservationTimeRepository;
+import roomescape.reservationTime.ReservationTime;
+import roomescape.reservationTime.ReservationTimeRepository;
 
 import java.util.Arrays;
 import java.util.List;
